results/2017-01-24/

+++ results/2016-12-19/README.sh
@@ -2,7 +2,7 @@
 
 #B.plicatilis.fa is the reference genome
 #B.plicatilis is the prefix for index files
-
+#maker2.all.gff is the structural annotation
 
 REFDIR=/data/eva/Brachionus/data/Transcriptoma/Genome
 DATA=/data/eva/Brachionus/data/Transcriptoma/Trim_reads
@@ -11,11 +11,11 @@ ANNOTATION=/data/eva/Brachionus/data/Transcriptoma/Annotation
 if [ ! -d index ]; then mkdir index; fi
 
 if [ ! -e index/B.plicatilis.fa ]; then
-   ln -s $REFDIR/B_plicat_PLATANUS.fasta index/B.plicatilis.fa
+   ln -s $REFDIR/Brachionus_plicatilis_scaffold_min500.fasta index/B.plicatilis.fa
 fi
 
 if [ ! -e index/B.plicatilis.1.bt2 ]; then
-   bowtie2-build index/B.plicatilis.fa index/B.plicatilis #Falta poner path (the recommendatio:all files in he same directory)
+   bowtie2-build index/B.plicatilis.fa index/B.plicatilis
 fi
 
 for i in $DATA/*.fastq.gz; do #Path para que vaya a los reads
@@ -29,8 +29,8 @@ for i in $DATA/*.fastq.gz; do #Path para que vaya a los reads
    fi
 
    #comprobar que options no dejar por defecto (ej. -p, etc)
-   #min-intron-length=10 (5 in positive strand and 15 in negative strand); max-intron-length=15000 (respect to negative strand) 	
-   if [ -e $name/accepted_hits.bam ]; then
+   #min-intron-length=10 (5 in positive strand and 15 in negative strand); max-intron-length=15000 (respect to negative strand)
+   if [ ! -e $name/accepted_hits.bam ]; then
       tophat2 -N 1 \
               --read-gap-length 1 \
               --read-realign-edit-dist 0 \
@@ -39,9 +39,10 @@ for i in $DATA/*.fastq.gz; do #Path para que vaya a los reads
               -I 15000 \
               --max-multihits 1 \
 	      -G $ANNOTATION/maker2.all.gff \
-              -o ./$name ./index/B.plicatilis $name.trim.fastq 1> $name/log 2> $name/error
-
-      rm $name.fastq
+              -o ./$name ./index/B.plicatilis $name.fastq 1> $name/log 2> $name/error
+      if [ -e $name/accepted_hits.bam ]; then
+         rm $name.fastq
+      fi
    fi
 done
